Fold ClientDirectStream and ServerDirectStream into RemoteEntryPoint
