+++ src/Workspaces/Remote/Razor/Microsoft.CodeAnalysis.Remote.Razor.ServiceHub.csproj
@@ -43,11 +43,5 @@
     <Compile Include="..\ServiceHub\Shared\RemoteEndPoint.cs">
       <Link>Shared\RemoteEndPoint.cs</Link>
     </Compile>
-    <Compile Include="..\ServiceHub\Shared\ClientDirectStream.cs">
-      <Link>Shared\ClientDirectStream.cs</Link>
-    </Compile>
-    <Compile Include="..\ServiceHub\Shared\ServerDirectStream.cs">
-      <Link>Shared\ServerDirectStream.cs</Link>
-    </Compile>
   </ItemGroup>
 </Project>
